--- src/jc21/Item.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace jc21;
-
-class Item
-{
-}
