+++ src/jc21/Util/Item.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace jc21\Util;
+
+interface Item
+{
+    public static function fromLibrary(array $lib): self;
+}
